feat: legacy details (#81)
* feat: legacy details

This adds a details blocks for legacy mode. If one is toggled, all are toggled. Alternative to #80

The idea is to add legacy syntax inline into the documentation at the appropriate places using these notes, and devs can hide those when unchecking the checkbox. The "Reactivity Fundamentals" page shows how this would look like.

* fix position when toggling other elements

* prevent flash-of-unopened-content

* tweaks

* tweak

* old habits die hard

* add show all/hide all text

* prettier

---------

Co-authored-by: Rich Harris <rich.harris@vercel.com>