docs: Improve explanation of 05-keyed-each-blocks in tutorial (#1252)
* Update index.md

I personally was very confused with the previous explanation as to what it meant that the other list items were updated. Explaining it that egg replaced the name attribute for the doughnut, and doughnut replaced the name attribute for the car node, helped me understand what Svelte was actually doing.

* Apply suggestions from code review

---------

Co-authored-by: Rich Harris <hello@rich-harris.dev>