fix: prevent the tutorial erroring after navigating (#1352)
* Fix the tutorial erroring after navigating

The `Workspace.reset` method was changed in a non-backward compatible
way in commit eb1eb2b0fd, which caused the call in the tutorial pages
to not pass the proper value for the selected file. This was missed
because the tutorial page didn't use `PageProps` for typing.

* Remove unneeded type annotation

The language server infers this automatically, turns out

* Update +page.svelte

---------

Co-authored-by: Tee Ming <chewteeming01@gmail.com>